synchronize with shell-skeleton

+++ script/skeleton/synchronize.sh
@@ -1,8 +1,12 @@
 #!/bin/sh -e
 
+DIRECTORY=$(dirname "${0}")
+SCRIPT_DIRECTORY=$(cd "${DIRECTORY}" || exit 1; pwd)
+# shellcheck source=/dev/null
+. "${SCRIPT_DIRECTORY}/../../lib/common.sh"
 TARGET="${1}"
 
-if [ "${TARGET}" = "" ]; then
+if [ "${TARGET}" = '' ]; then
     echo "Usage: ${0} TARGET"
 
     exit 1
@@ -14,9 +18,9 @@ if [ ! -d "${TARGET}" ]; then
     exit 1
 fi
 
-NAME=$(head -n1 "${TARGET}/README.md" | awk '{ print $2 }' | grep --extended-regexp '^([A-Z]+[a-z0-9]*){2,}$') || NAME=""
+NAME=$(head -n 1 "${TARGET}/README.md" | awk '{ print $2 }' | grep --extended-regexp '^([A-Z]+[a-z0-9]*){1,}$') || NAME=''
 
-if [ "${NAME}" = "" ]; then
+if [ "${NAME}" = '' ]; then
     echo "Could not determine the project name."
 
     exit 1
@@ -37,18 +41,25 @@ mkdir -p "${TARGET}/documentation"
 cp -R documentation/* "${TARGET}/documentation"
 mkdir -p "${TARGET}/script"
 cp -R script/* "${TARGET}/script"
+mkdir -p "${TARGET}/lib"
+cp lib/common.sh "${TARGET}/lib"
+cp .gitignore "${TARGET}"
+cp Vagrantfile "${TARGET}"
 cp sonar-project.properties "${TARGET}"
 cp composer.json "${TARGET}"
 cp phpunit.xml "${TARGET}"
-cp .gitignore "${TARGET}"
 cp .php_cs.php "${TARGET}"
 cp .phpmd.xml "${TARGET}"
 cp .phpunit.ci.xml "${TARGET}"
 cp .phpstan.neon "${TARGET}"
-cp Vagrantfile "${TARGET}"
 cd "${TARGET}" || exit 1
-rm -rf script/skeleton
+echo "${NAME}" | grep --quiet 'Skeleton$' && IS_SKELETON=true || IS_SKELETON=false
+
+if [ "${IS_SKELETON}" = false ]; then
+    rm -rf script/skeleton
+fi
+
 DASH=$(echo "${NAME}" | ${SED} --regexp-extended 's/([A-Za-z0-9])([A-Z])/\1-\2/g' | tr '[:upper:]' '[:lower:]')
 INITIALS=$(echo "${NAME}" | ${SED} 's/\([A-Z]\)[a-z]*/\1/g' | tr '[:upper:]' '[:lower:]')
 # shellcheck disable=SC2016
-${FIND} . -type f -regextype posix-extended ! -regex '^.*/(build|vendor|\.git|\.idea)/.*$' -exec sh -c '${1} --in-place --expression "s/PhpSkeleton/${2}/g" --expression "s/php-skeleton/${3}/g" --expression "s/bin\/ps/bin\/${4}/g" "${5}"' '_' "${SED}" "${NAME}" "${DASH}" "${INITIALS}" '{}' \;
+${FIND} . -regextype posix-extended -type f ! -regex "${EXCLUDE_FILTER}" -exec sh -c '${1} --in-place --expression "s/PhpSkeleton/${2}/g" --expression "s/php-skeleton/${3}/g" --expression "s/bin\/ss/bin\/${4}/g" --expression "s/ps\\\\/${4}\\\\/g" "${5}"' '_' "${SED}" "${NAME}" "${DASH}" "${INITIALS}" '{}' \;
